Changes in script and included classes.
The script included all xsd sub-folders, but the examples naively referenced classes without any prefixed.
For GML we now have explicitly added them (manually) to the script, with the gml: prefix. The examples were
updated accordingly.

This significantly reduces the enumerations, and reduces them to NeTEx only (without prefix).

In addition substitutionGroups 'results' have been added to the enumeration of the classes.